refactor: error types
