+++ lib/errors/types/not-found.ts
