+++ lib/errors/types/reject.ts
