+++ lib/errors/types/request-in-progress.ts
