+++ lib/middleware/access-control.ts
@@ -1,7 +1,7 @@
 import type { MiddlewareHandler } from 'hono';
 import { config } from '@/config';
 import md5 from '@/utils/md5';
-import RejectError from '@/errors/reject';
+import RejectError from '@/errors/types/reject';
 
 const reject = () => {
     throw new RejectError('Authentication failed. Access denied.');
